snowbridge_inbound_queue_primitives::v2::traits

Trait MessageProcessor

pub trait MessageProcessor<AccountId> {
    // Required methods
    fn can_process_message(relayer: &AccountId, message: &Message) -> bool;
    fn process_message(
        relayer: AccountId,
        message: Message,
    ) -> Result<[u8; 32], MessageProcessorError>;
}

Trait to define the logic for checking and processing inbound messages.

Required Methods

fn can_process_message(relayer: &AccountId, message: &Message) -> bool

Lightweight function to check if this processor can handle the message

fn process_message( relayer: AccountId, message: Message, ) -> Result<[u8; 32], MessageProcessorError>

Process the message and return the message ID

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.